# openapi.yaml
openapi: 3.0.3
info:
  title: Pessoa Controler API
  description: Isto é a descricao da controller Pessoa API
  termsOfService: http://swagger.io/terms/
  contact:
    email: mbsoares.infor@gmail.com
  version: 1.0.11
...
paths:
  /pessoa:
    post:
      tags:
        - crud pessoa
      summary: Cadastro de pessoa
      description: Cadastramento de Pessoa
      operationId: salvarPessoa
      requestBody:
        description: Entidade/Modelo/Schema Pessoa
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pessoa'
        required: true
      responses:
        '400':
          description: Successful operation
    put:
      summary: Atualização de Pessoa
      description: Atualiza uma pessoa
      operationId: atualizarPessoa
      requestBody:
        description: Entidade/Modelo/Schema Pessoa
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pessoa'
        required: true
      responses:
        '400':
          description: Successful operation
      tags:
        - crud pessoa
...
components:
  schemas:
    Pessoa:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 5
        nome:
          type: string
          example: pedro
        endereco:
          type: string
          example: teste